# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "to-do-Movie API",
        "termsOfService": "http://localhost:8081/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1/",
    "paths": {
        "/movies": {
            "get": {
                "description": "show all movies",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GET All Movies"
                ],
                "summary": "show all movies",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/server_side.Movie"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server_side.Movie"
                        }
                    }
                }
            },
            "post": {
                "description": "create movie",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "POST Movie"
                ],
                "summary": "create movie",
                "responses": {
                    "200": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/movies/id/{id}": {
            "get": {
                "description": "show movie by Id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GET Movie by Id"
                ],
                "summary": "show movie by Id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server_side.Movie"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server_side.Movie"
                        }
                    }
                }
            },
            "put": {
                "description": "update movie by Id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UPDATE Movie by Id"
                ],
                "summary": "update movie by Id",
                "responses": {
                    "200": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "No Content"
                    }
                }
            },
            "delete": {
                "description": "delete movie by Id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DELETE Movie by Id"
                ],
                "summary": "delete movie by Id",
                "responses": {
                    "200": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "No Content"
                    }
                }
            }
        }
    },
    "definitions": {
        "server_side.Movie": {
            "type": "object",
            "properties": {
                "director": {
                    "type": "string"
                },
                "done": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        }
    }
}